Create README.md
# RECIPES
# PROJECT DESCRIPTION
This is a project from which you can search for different food recipes by categories, see the detail of the same by clicking on the image and also allows you to scroll horizontally to see the different recipes with which the application counts.

## Necessary tool:
Install node.js

## Steps to execute:
Step 1. Download or Clone the repository Step 2. Unzip the file if you have downloaded it Step 3. Delete the file named 'package-lock.json' Step 4. Install npm with the finish 'npm install' Step 5. Type in the terminal the command 'npx react-native run-android' to run the project.

## Author
Tony Vasquez Arias